chore(trellis): updated Trellis to the latest version
In preparation for migrating to a new server, we’re bumping Trellis
to the latest version. Should fix a Let’s Encrypt config issue and
also (hopefully) eliminate the issue where Nginx is caching its own
“ready” page instead of server the actual site.

However, I am not good at devops. So this is kind of a “spray and
pray” scenario. Here goes nothing.